remove global cache lock and expensive operations for statistics access,  https://github.com/cache2k/cache2k/issues/150

+++ cache2k-core/src/main/java/org/cache2k/core/ConcurrentMapWrapper.java
@@ -108,7 +108,7 @@ public class ConcurrentMapWrapper<K, V> extends AbstractMap<K, V> implements Con
 
   @Override
   public int size() {
-    return ((InternalCache<?, ?>) cache).getTotalEntryCount();
+    return (int) ((InternalCache<?, ?>) cache).getTotalEntryCount();
   }
 
   @Override
